+++ Packages.props
@@ -1,34 +1,34 @@
 <Project>
 
   <ItemGroup>
-    <PackageReference Update="AvantiPoint.CodeGenHelpers" Version="1.4.9" />
-    <PackageReference Update="Microsoft.Build" Version="16.0.461" />
-    <PackageReference Update="Microsoft.Build.Tasks.Core" Version="16.0.461" />
-    <PackageReference Update="Microsoft.CodeAnalysis.Common" Version="3.11.0" />
+    <PackageReference Update="AvantiPoint.CodeGenHelpers" Version="1.5.6" />
+    <PackageReference Update="Microsoft.Build" Version="17.0.0" />
+    <PackageReference Update="Microsoft.Build.Tasks.Core" Version="17.0.0" />
+    <PackageReference Update="Microsoft.CodeAnalysis.Common" Version="4.2.0" />
     <PackageReference Update="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Update="Newtonsoft.Json.Schema" Version="3.0.14" />
     <PackageReference Update="NuGet.Build.Packaging" Version="0.2.5-dev.10" />
-    <PackageReference Update="NuGetizer" Version="0.7.5" />
+    <PackageReference Update="NuGetizer" Version="0.8.0" />
     <PackageReference Update="Microsoft.Web.Xdt" Version="3.1.0" />
 
     <PackageReference Update="Fizzler" Version="1.3.0" />
-    <PackageReference Update="HarfBuzzSharp" Version="2.6.1.7" />
-    <PackageReference Update="ShimSkiaSharp" Version="0.5.5.5" />
-    <PackageReference Update="SkiaSharp" Version="2.80.2" />
+    <PackageReference Update="HarfBuzzSharp" Version="2.8.2" />
+    <PackageReference Update="ShimSkiaSharp" Version="0.5.14" />
+    <PackageReference Update="SkiaSharp" Version="2.88.0" />
     <PackageReference Update="SkiaSharp.HarfBuzz" Version="2.80.2" />
-    <PackageReference Update="Svg.Skia" Version="0.5.5.5" />
-    <PackageReference Update="Svg.Custom" Version="0.5.5.5" />
-    <PackageReference Update="Svg.Model" Version="0.5.5.5" />
+    <PackageReference Update="Svg.Skia" Version="0.5.14" />
+    <PackageReference Update="Svg.Custom" Version="0.5.14" />
+    <PackageReference Update="Svg.Model" Version="0.5.14" />
     <PackageReference Update="System.Buffers" Version="4.5.1" />
     <PackageReference Update="System.ComponentModel.Annotations" Version="5.0.0" />
     <PackageReference Update="System.Drawing.Common" Version="5.0.2" />
     <PackageReference Update="System.IO.UnmanagedMemoryStream" Version="4.3.0" />
-    <PackageReference Update="System.Memory" Version="4.5.4" />
+    <PackageReference Update="System.Memory" Version="4.5.5" />
     <PackageReference Update="System.Numerics.Vectors" Version="4.5.0" />
     <PackageReference Update="System.ObjectModel" Version="4.3.0" />
-    <PackageReference Update="System.Runtime.CompilerServices.Unsafe" Version="5.0.0" />
-    <PackageReference Update="System.Text.Json" Version="5.0.0" />
-    <PackageReference Update="McMaster.Extensions.CommandLineUtils" Version="3.1.0" />
+    <PackageReference Update="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
+    <PackageReference Update="System.Text.Json" Version="6.0.5" />
+    <PackageReference Update="McMaster.Extensions.CommandLineUtils" Version="4.0.1" />
   </ItemGroup>
 
   <ItemGroup>
@@ -40,10 +40,10 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Update="Microsoft.NET.Test.Sdk" Version="16.10.0" />
-    <PackageReference Update="Moq" Version="4.16.1" />
+    <PackageReference Update="Microsoft.NET.Test.Sdk" Version="17.2.0" />
+    <PackageReference Update="Moq" Version="4.18.1" />
     <PackageReference Update="xunit" Version="2.4.1" />
-    <PackageReference Update="xunit.runner.visualstudio" Version="2.4.3" />
+    <PackageReference Update="xunit.runner.visualstudio" Version="2.4.5" />
     <PackageReference Update="NUnit" Version="3.13.1" />
     <PackageReference Update="Xamarin.UITest" Version="3.0.14" />
   </ItemGroup>
