+++ global.json
@@ -1,11 +1,11 @@
 {
   "sdk": {
-    "version": "6.0.300",
+    "version": "6.0.301",
     "rollForward": "latestMinor",
     "allowPrerelease": false
   },
   "msbuild-sdks": {
-    "MSBuild.Sdk.Extras": "3.0.23",
-    "Microsoft.Build.NoTargets": "3.3.0"
+    "MSBuild.Sdk.Extras": "3.0.44",
+    "Microsoft.Build.NoTargets": "3.5.6"
   }
 }
