+++ src/Mobile.BuildTools.Images/Android.targets
@@ -1,0 +1,14 @@
+<Project>
+
+  <Target Name="IncludeUnifiedImageAsset"
+          AfterTargets="UnifiedImageAssetProcessing"
+          BeforeTargets="UpdateAndroidResources"
+          Condition=" $(BuildToolsBundleUnifiedImages) == 'true' ">
+    <ItemGroup>
+      <AndroidResource Include="@(UnifiedImageAsset)"
+                       Link="%(UnifiedImageAsset.OutputLink)"
+                       Visible="%(UnifiedImageAsset.ShouldBeVisible)" />
+    </ItemGroup>
+  </Target>
+
+</Project>
