+++ src/Mobile.BuildTools.Images/Drawing/Context.cs
