+++ src/Mobile.BuildTools.Images/Drawing/Image.cs
