+++ src/Mobile.BuildTools.Images/Drawing/ImageBase.cs
