+++ src/Mobile.BuildTools.Images/Drawing/SKBitmapExtensions.cs
