+++ src/Mobile.BuildTools.Images/Models/AppIcons/OutputImage.partial.cs
