+++ src/Mobile.BuildTools/UAP.targets
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <Project>
 
-  <Import Project="ImageResize.targets" />
   <Import Project="ConfigurationManager.targets" />
   <!-- When the Manifests are supported use the PlatformHead.targets
   <Import Project="ManifestsShared.targets" />
@@ -14,16 +13,6 @@
     <ItemGroup>
       <Content Include="@(GeneratedAppConfig)"
                       Link="Assets\%(Filename)%(Extension)" />
-    </ItemGroup>
-  </Target>
-
-  <Target Name="IncludeUnifiedImageAsset"
-          BeforeTargets="_CollectBundleResources"
-          AfterTargets="UnifiedImageAssetProcessing;_CollectImageAssets">
-    <ItemGroup>
-      <Content Include="@(UnifiedImageAsset)"
-               Link="%(UnifiedImageAsset.OutputLink)"
-               Visible="true" />
     </ItemGroup>
   </Target>
