+++ src/Mobile.BuildTools/iOS.targets
@@ -25,22 +25,6 @@
     </ItemGroup>
   </Target>
 
-  <Target Name="IncludeUnifiedImageAsset"
-          BeforeTargets="_CollectBundleResources"
-          AfterTargets="UnifiedImageAssetProcessing;_CollectImageAssets">
-    <ItemGroup>
-      <BundleResource Include="@(UnifiedImageAsset)"
-                      Link="%(UnifiedImageAsset.OutputLink)"
-                      Visible="true"
-                      Condition="%(UnifiedImageAsset.ShouldBeVisible) == 'true' " />
-
-      <ImageAsset Include="@(UnifiedImageAsset)"
-                  Link="%(UnifiedImageAsset.OutputLink)"
-                  Visible="false"
-                  Condition="%(UnifiedImageAsset.ShouldBeVisible) != 'true' " />
-    </ItemGroup>
-  </Target>
-
   <Target Name="IPACopyToStagingDirectory"
           AfterTargets="_ZipIpa"
           DependsOnTargets="_ZipIpa"
